% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snacks.R
\name{snacks}
\alias{snacks}
\title{Title}
\usage{
snacks(grapes, apples, burritos)
}
\arguments{
\item{grapes}{mass of grapes}

\item{apples}{mass of apples}

\item{burritos}{how many burritos I eat}
}
\description{
Title
}
\examples{
snacks(grapes = 2.5, apples = 6, burritos = 1)
}
