% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekday.R
\name{weekday}
\alias{weekday}
\title{Title}
\usage{
weekday(weekday, name)
}
\arguments{
\item{weekday}{any day of the week}

\item{name}{any name}
}
\description{
Title
}
\examples{
weekday("saturday","suzie")
}
